TF-2520 Create some class to handle html
