+++ core/lib/core.dart
@@ -1,101 +1,10 @@
 library core;
 
-// Extensions
-export 'presentation/extensions/color_extension.dart';
-export 'presentation/extensions/url_extension.dart';
-export 'presentation/extensions/uri_extension.dart';
-export 'presentation/extensions/capitalize_extension.dart';
-export 'presentation/extensions/list_extensions.dart';
-export 'domain/extensions/datetime_extension.dart';
-export 'presentation/extensions/html_extension.dart';
-export 'presentation/extensions/compare_string_extension.dart';
-export 'presentation/extensions/compare_list_extensions.dart';
-export 'presentation/extensions/string_extension.dart';
-export 'presentation/extensions/tap_down_details_extension.dart';
-export 'domain/extensions/media_type_extension.dart';
-export 'presentation/extensions/map_extensions.dart';
-
-// Exceptions
-export 'domain/exceptions/download_file_exception.dart';
-export 'data/extensions/options_extensions.dart';
-export 'domain/exceptions/web_session_exception.dart';
-
-// Utils
-export 'presentation/utils/theme_utils.dart';
-export 'presentation/utils/responsive_utils.dart';
-export 'presentation/utils/keyboard_utils.dart';
-export 'presentation/utils/icon_utils.dart';
-export 'presentation/utils/style_utils.dart';
-export 'presentation/utils/app_toast.dart';
-export 'presentation/utils/html_transformer/html_template.dart';
-export 'presentation/utils/html_transformer/html_transform.dart';
-export 'presentation/utils/html_transformer/transform_configuration.dart';
-export 'presentation/utils/html_transformer/dom/add_tooltip_link_transformers.dart';
-export 'data/utils/device_manager.dart';
-export 'utils/app_logger.dart';
-export 'utils/benchmark.dart';
-export 'utils/fps_manager.dart';
-export 'utils/build_utils.dart';
-export 'utils/string_convert.dart';
-export 'utils/config/app_config_loader.dart';
-export 'utils/config/app_config_parser.dart';
-export 'utils/config/errors.dart';
-export 'data/utils/compress_file_utils.dart';
-export 'utils/platform_info.dart';
-export 'utils/option_param_mixin.dart';
-export 'utils/file_utils.dart';
-export 'utils/print_utils.dart';
-
-// Views
-export 'presentation/views/text/slogan_builder.dart';
-export 'presentation/views/text/text_field_builder.dart';
-export 'presentation/views/text/text_form_field_builder.dart';
-export 'presentation/views/text/input_decoration_builder.dart';
-export 'presentation/views/text/rich_text_builder.dart';
-export 'presentation/views/text/text_overflow_builder.dart';
-export 'presentation/views/responsive/responsive_widget.dart';
-export 'presentation/views/list/tree_view.dart';
-export 'presentation/views/button/icon_button_web.dart';
-export 'presentation/views/button/tmail_button_widget.dart';
-export 'presentation/views/image/avatar_builder.dart';
-export 'presentation/views/list/sliver_grid_delegate_fixed_height.dart';
-export 'presentation/views/image/icon_builder.dart';
-export 'presentation/views/context_menu/context_menu_action_builder.dart';
-export 'presentation/views/context_menu/context_menu_builder.dart';
-export 'presentation/views/context_menu/context_menu_header_builder.dart';
-export 'presentation/views/context_menu/simple_context_menu_action_builder.dart';
-export 'presentation/views/dialog/downloading_file_dialog_builder.dart';
-export 'presentation/views/dialog/confirmation_dialog_builder.dart';
-export 'presentation/views/dialog/edit_text_dialog_builder.dart';
-export 'presentation/views/dialog/color_picker_dialog_builder.dart';
-export 'presentation/views/html_viewer/html_content_viewer_widget.dart';
-export 'presentation/views/html_viewer/html_content_viewer_on_web_widget.dart';
-export 'presentation/views/floating_button/scrolling_floating_button_animated.dart';
-export 'presentation/views/bottom_popup/cupertino_action_sheet_action_builder.dart';
-export 'presentation/views/bottom_popup/cupertino_action_sheet_no_icon_builder.dart';
-export 'presentation/views/bottom_popup/cupertino_action_sheet_builder.dart';
-export 'presentation/views/bottom_popup/confirmation_dialog_action_sheet_builder.dart';
-export 'presentation/views/modal_sheets/edit_text_modal_sheet_builder.dart';
-export 'presentation/views/search/search_bar_view.dart';
-export 'presentation/views/popup_menu/popup_menu_item_widget.dart';
-export 'presentation/views/quick_search/quick_search_input_form.dart';
-export 'presentation/views/toast/toast_position.dart';
-export 'presentation/views/toast/tmail_toast.dart';
-export 'presentation/views/bottom_popup/full_screen_action_sheet_builder.dart';
-export 'presentation/views/checkbox/labeled_checkbox.dart';
-export 'presentation/views/container/tmail_container_widget.dart';
-export 'presentation/views/clipper/side_arrow_clipper.dart';
-export 'presentation/views/avatar/gradient_circle_avatar_icon.dart';
-export 'presentation/views/loading/cupertino_loading_widget.dart';
-
-// Resources
-export 'presentation/resources/assets_paths.dart';
-export 'presentation/resources/image_paths.dart';
-
-// Constants
-export 'presentation/constants/constants_ui.dart';
 export 'data/constants/constant.dart';
-
+export 'data/extensions/options_extensions.dart';
+export 'data/model/query/query_parameter.dart';
+// Model
+export 'data/model/source_type/data_source_type.dart';
 // Network
 export 'data/network/config/dynamic_url_interceptors.dart';
 export 'data/network/config/service_path.dart';
@@ -103,14 +12,96 @@ export 'data/network/dio_client.dart';
 export 'data/network/download/download_client.dart';
 export 'data/network/download/download_manager.dart';
 export 'data/network/download/downloaded_response.dart';
-
-// State
-export 'presentation/state/success.dart';
-export 'presentation/state/failure.dart';
-
-// Model
-export 'data/model/source_type/data_source_type.dart';
-export 'data/model/query/query_parameter.dart';
-
+export 'data/utils/compress_file_utils.dart';
+export 'data/utils/device_manager.dart';
+// Exceptions
+export 'domain/exceptions/download_file_exception.dart';
+export 'domain/exceptions/web_session_exception.dart';
+export 'domain/extensions/datetime_extension.dart';
+export 'domain/extensions/media_type_extension.dart';
 // Action
 export 'presentation/action/action_callback_define.dart';
+// Constants
+export 'presentation/constants/constants_ui.dart';
+export 'presentation/extensions/capitalize_extension.dart';
+// Extensions
+export 'presentation/extensions/color_extension.dart';
+export 'presentation/extensions/compare_list_extensions.dart';
+export 'presentation/extensions/compare_string_extension.dart';
+export 'presentation/extensions/html_extension.dart';
+export 'presentation/extensions/list_extensions.dart';
+export 'presentation/extensions/map_extensions.dart';
+export 'presentation/extensions/string_extension.dart';
+export 'presentation/extensions/tap_down_details_extension.dart';
+export 'presentation/extensions/uri_extension.dart';
+export 'presentation/extensions/url_extension.dart';
+// Resources
+export 'presentation/resources/assets_paths.dart';
+export 'presentation/resources/image_paths.dart';
+export 'presentation/state/failure.dart';
+// State
+export 'presentation/state/success.dart';
+export 'presentation/utils/app_toast.dart';
+export 'presentation/utils/html_transformer/dom/add_tooltip_link_transformers.dart';
+export 'presentation/utils/html_transformer/html_transform.dart';
+export 'presentation/utils/html_transformer/transform_configuration.dart';
+export 'presentation/utils/icon_utils.dart';
+export 'presentation/utils/keyboard_utils.dart';
+export 'presentation/utils/responsive_utils.dart';
+export 'presentation/utils/style_utils.dart';
+// Utils
+export 'presentation/utils/theme_utils.dart';
+export 'presentation/views/avatar/gradient_circle_avatar_icon.dart';
+export 'presentation/views/bottom_popup/confirmation_dialog_action_sheet_builder.dart';
+export 'presentation/views/bottom_popup/cupertino_action_sheet_action_builder.dart';
+export 'presentation/views/bottom_popup/cupertino_action_sheet_builder.dart';
+export 'presentation/views/bottom_popup/cupertino_action_sheet_no_icon_builder.dart';
+export 'presentation/views/bottom_popup/full_screen_action_sheet_builder.dart';
+export 'presentation/views/button/icon_button_web.dart';
+export 'presentation/views/button/tmail_button_widget.dart';
+export 'presentation/views/checkbox/labeled_checkbox.dart';
+export 'presentation/views/clipper/side_arrow_clipper.dart';
+export 'presentation/views/container/tmail_container_widget.dart';
+export 'presentation/views/context_menu/context_menu_action_builder.dart';
+export 'presentation/views/context_menu/context_menu_builder.dart';
+export 'presentation/views/context_menu/context_menu_header_builder.dart';
+export 'presentation/views/context_menu/simple_context_menu_action_builder.dart';
+export 'presentation/views/dialog/color_picker_dialog_builder.dart';
+export 'presentation/views/dialog/confirmation_dialog_builder.dart';
+export 'presentation/views/dialog/downloading_file_dialog_builder.dart';
+export 'presentation/views/dialog/edit_text_dialog_builder.dart';
+export 'presentation/views/floating_button/scrolling_floating_button_animated.dart';
+export 'presentation/views/html_viewer/html_content_viewer_on_web_widget.dart';
+export 'presentation/views/html_viewer/html_content_viewer_widget.dart';
+export 'presentation/views/image/avatar_builder.dart';
+export 'presentation/views/image/icon_builder.dart';
+export 'presentation/views/list/sliver_grid_delegate_fixed_height.dart';
+export 'presentation/views/list/tree_view.dart';
+export 'presentation/views/loading/cupertino_loading_widget.dart';
+export 'presentation/views/modal_sheets/edit_text_modal_sheet_builder.dart';
+export 'presentation/views/popup_menu/popup_menu_item_widget.dart';
+export 'presentation/views/quick_search/quick_search_input_form.dart';
+export 'presentation/views/responsive/responsive_widget.dart';
+export 'presentation/views/search/search_bar_view.dart';
+export 'presentation/views/text/input_decoration_builder.dart';
+export 'presentation/views/text/rich_text_builder.dart';
+// Views
+export 'presentation/views/text/slogan_builder.dart';
+export 'presentation/views/text/text_field_builder.dart';
+export 'presentation/views/text/text_form_field_builder.dart';
+export 'presentation/views/text/text_overflow_builder.dart';
+export 'presentation/views/toast/tmail_toast.dart';
+export 'presentation/views/toast/toast_position.dart';
+export 'utils/app_logger.dart';
+export 'utils/benchmark.dart';
+export 'utils/build_utils.dart';
+export 'utils/config/app_config_loader.dart';
+export 'utils/config/app_config_parser.dart';
+export 'utils/config/errors.dart';
+export 'utils/file_utils.dart';
+export 'utils/fps_manager.dart';
+export 'utils/html/html_template.dart';
+export 'utils/option_param_mixin.dart';
+export 'utils/platform_info.dart';
+export 'utils/print_utils.dart';
+export 'utils/string_convert.dart';
