+++ core/lib/presentation/utils/html_transformer/dom/remove_tooltip_link_transformers.dart
@@ -1,8 +1,8 @@
 
 import 'package:core/data/network/dio_client.dart';
 import 'package:core/presentation/utils/html_transformer/base/dom_transformer.dart';
-import 'package:core/presentation/utils/html_transformer/html_template.dart';
 import 'package:core/utils/app_logger.dart';
+import 'package:core/utils/html/html_template.dart';
 import 'package:html/dom.dart';
 
 class RemoveTooltipLinkTransformer extends DomTransformer {
@@ -15,11 +15,11 @@ class RemoveTooltipLinkTransformer extends DomTransformer {
     required DioClient dioClient,
     Map<String, String>? mapUrlDownloadCID,
   }) async {
-    final linkElements = document.querySelectorAll('a.$nameClassToolTip');
+    final linkElements = document.querySelectorAll('a.${HtmlTemplate.nameClassToolTip}');
     await Future.wait(linkElements.map((linkElement) async {
       final classAttribute = linkElement.attributes['class'];
       if (classAttribute != null) {
-        final newClassAttribute = classAttribute.replaceFirst(nameClassToolTip, '');
+        final newClassAttribute = classAttribute.replaceFirst(HtmlTemplate.nameClassToolTip, '');
         linkElement.attributes['class'] = newClassAttribute;
       }
       final listSpanTag = linkElement.querySelectorAll('span.tooltiptext');
