--- core/lib/presentation/utils/html_transformer/html_event_action.dart
@@ -1,5 +1,0 @@
-
-class HtmlEventAction {
-  static const scrollRightEndAction = 'ScrollRightEndAction';
-  static const scrollLeftEndAction = 'ScrollLeftEndAction';
-}
